+++ Skyflow/src/main/kotlin/Skyflow/BaseElement.kt
@@ -4,9 +4,7 @@ import android.content.Context
 import android.util.AttributeSet
 import android.widget.LinearLayout
 
-abstract class BaseElement @JvmOverloads constructor(
-    context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0,
-)  : LinearLayout(context, attrs, defStyleAttr){
-
-      internal abstract fun getValue():String
+interface  BaseElement   {
+       fun setError(error:String)
+       fun resetError()
  }