+++ Skyflow/src/main/kotlin/Skyflow/CollectOptions.kt
@@ -2,5 +2,5 @@ package Skyflow
 
 import org.json.JSONObject
 
-class CollectOptions(val token:Boolean = true,val additionalFields : JSONObject = JSONObject()) {
+class CollectOptions(val token:Boolean = true, val additionalFields: JSONObject? = null) {
 }