+++ Skyflow/src/main/kotlin/Skyflow/Configuration.kt
@@ -3,8 +3,8 @@ package Skyflow
 
 
 class Configuration(
-    val vaultID: String,
-    var vaultURL: String,
+    val vaultID: String = "",
+    var vaultURL: String = "",
     val tokenProvider: TokenProvider,
     val options: Options = Options(),
 ){