+++ Skyflow/src/main/kotlin/Skyflow/ElementType.kt
@@ -4,7 +4,7 @@ import com.Skyflow.collect.elements.validations.*
 import com.Skyflow.collect.elements.validations.LengthMatchRule
 import com.Skyflow.collect.elements.validations.RegexMatchRule
 import com.Skyflow.collect.elements.validations.SkyflowValidateCardNumber
-import com.Skyflow.collect.elements.validations.SkyflowValidateExpirationDate
+import com.Skyflow.collect.elements.validations.SkyflowValidateExpireDate
 import com.Skyflow.collect.elements.validations.SkyflowValidateLengthMatch
 import com.Skyflow.collect.elements.validations.SkyflowValidationErrorType
 
@@ -52,7 +52,7 @@ enum class SkyflowElementType {
                 return Type(
                     "#### #### #### ####",
                     "^(?:4[0-9]{12}(?:[0-9]{3})?|[25][1-7][0-9]{14}|6(?:011|5[0-9][0-9])[0-9]{12}|3[47][0-9]{13}|3(?:0[0-5]|[68][0-9])[0-9]{11}|(?:2131|1800|35\\d{3})\\d{11})$",
-                    rules, InputType.TYPE_CLASS_TEXT
+                    rules, InputType.TYPE_CLASS_NUMBER
                 )
             }
             CVV -> {
@@ -66,7 +66,7 @@ enum class SkyflowElementType {
                 )
                 return Type(
                     "####", "\\d*$",
-                    rules, InputType.TYPE_CLASS_TEXT
+                    rules, InputType.TYPE_CLASS_NUMBER
                 )
             }
             EXPIRATION_DATE -> {
@@ -74,14 +74,9 @@ enum class SkyflowElementType {
                     RegexMatchRule("^(0[1-9]|1[0-2])\\/?([0-9]{4}|[0-9]{2})$",
                     SkyflowValidationErrorType.pattern.rawValue)
                 )
-                rules.add(
-                    SkyflowValidateExpirationDate(
-                        SkyflowValidationErrorType.expireDate.rawValue
-                    )
-                )
                 return Type(
                     "##/##", "^(0[1-9]|1[0-2])\\/?([0-9]{4}|[0-9]{2})$",
-                    rules, InputType.TYPE_CLASS_TEXT
+                    rules, InputType.TYPE_CLASS_DATETIME
                 )
             }
             INPUT_FIELD  -> {