+++ Skyflow/src/main/kotlin/Skyflow/Label.kt
@@ -14,7 +14,7 @@ import androidx.core.content.res.ResourcesCompat
 @Suppress("DEPRECATION")
 class Label @JvmOverloads constructor(
     context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0,
-) : BaseElement(context, attrs, defStyleAttr) {
+) : LinearLayout(context, attrs, defStyleAttr),BaseElement {
 
     internal  var actualValue: String =""
     internal var label = TextView(context)
@@ -25,6 +25,7 @@ class Label @JvmOverloads constructor(
     internal lateinit var padding: Padding
     internal var border = GradientDrawable()
     internal var isTokenNull = false
+    internal var isError = false
 
     @SuppressLint("NewApi", "WrongConstant")
     internal fun setupField(revealInput: RevealElementInput, options: RevealElementOptions)
@@ -32,7 +33,7 @@ class Label @JvmOverloads constructor(
         this.revealInput = revealInput
         this.options = options
         padding = revealInput.inputStyles.base.padding
-        if(this.revealInput.token.equals("null") || this.revealInput.token.equals(null))
+        if(this.revealInput.token.equals(null))
         {
             isTokenNull = true
             this.revealInput.token = ""
@@ -54,13 +55,7 @@ class Label @JvmOverloads constructor(
     }
 
     private fun buildPlaceholder() {
-        if(revealInput.altText.isEmpty() || revealInput.altText == "")
-        {
-            placeholder.text = revealInput.token
-        }
-        else
-            placeholder.text = revealInput.altText
-
+        setText()
         if(revealInput.inputStyles.base.font != Typeface.NORMAL)
              placeholder.typeface = ResourcesCompat.getFont(context,revealInput.inputStyles.base.font)
         placeholder.textSize = 20f
@@ -93,9 +88,100 @@ class Label @JvmOverloads constructor(
         addView(error)
     }
 
-    override fun getValue(): String {
-        return actualValue
+    fun setErrorText(error:String)
+    {
+        this.error.text = error
     }
 
+    internal fun getValue(): String {
+        return actualValue
+    }
+    override fun setError(error: String) {
+        isError = true
+        setErrorText(error)
+        showError()
+    }
+
+    override fun resetError() {
+        isError = false
+        hideError()
+    }
+
+    internal fun showError(){
+        setErrorStyles()
+        setInvalidStyles()
+        this.error.visibility = VISIBLE
+    }
+
+    private fun hideError() {
+        setErrorText("")
+        this.error.visibility = INVISIBLE
+        buildPlaceholder()
+        buildLabel()
+    }
+
+    private fun setInvalidStyles(){
+        if(this.revealInput.inputStyles.invalid.font != Typeface.NORMAL)
+            this.placeholder.typeface =
+                ResourcesCompat.getFont(this.context,
+                    this.revealInput.inputStyles.invalid.font)
+        this.placeholder.gravity =
+            this.revealInput.inputStyles.invalid.textAlignment
+        val padding = this.revealInput.inputStyles.invalid.padding
+        this.placeholder.setPadding(padding.left,
+            padding.top,
+            padding.right,
+            padding.bottom)
+        this.placeholder.setTextColor(this.revealInput.inputStyles.invalid.textColor)
+        this.border.setStroke(this.revealInput.inputStyles.invalid.borderWidth,
+            this.revealInput.inputStyles.invalid.borderColor)
+        this.border.cornerRadius =
+            this.revealInput.inputStyles.invalid.cornerRadius
+        this.placeholder.setBackgroundDrawable(this.border)
+    }
+
+    private fun setErrorStyles(){
+        error.textSize = 16F
+        val errorPadding = revealInput.errorTextStyles.base.padding
+        error.setPadding(errorPadding.left,errorPadding.top,errorPadding.right,errorPadding.bottom)
+        error.setTextColor(revealInput.errorTextStyles.base.textColor)
+        if(revealInput.errorTextStyles.base.font != Typeface.NORMAL)
+            error.typeface = ResourcesCompat.getFont(context,revealInput.errorTextStyles.base.font)
+        error.gravity = revealInput.errorTextStyles.base.textAlignment
+    }
+
+    internal fun getErrorText() :String
+    {
+        return this.error.text.toString()
+    }
+
+    fun setToken(token : String)
+    {
+        this.revealInput.token = token
+        setText()
+    }
+    fun setAltText(altText:String)
+    {
+        this.revealInput.altText = altText
+        setText()
+
+    }
+
+    fun clearAltText()
+    {
+        this.revealInput.altText = ""
+        if(this.actualValue.isEmpty())
+            setText()
+    }
+
+    private fun setText()
+    {
+        if(revealInput.altText.isEmpty() || revealInput.altText == "")
+        {
+            placeholder.text = revealInput.token
+        }
+        else
+            placeholder.text = revealInput.altText
+    }
 
 }