+++ Skyflow/src/main/kotlin/Skyflow/SkyflowError.kt
@@ -5,17 +5,19 @@ import Skyflow.utils.Utils
 
 
 
-class SkyflowError(val skyflowErrorCode: SkyflowErrorCode = SkyflowErrorCode.UNKNOWN_ERROR, val tag : String? = "", logLevel: LogLevel = LogLevel.ERROR, params: Array<String?> = arrayOf()) : Exception(skyflowErrorCode.getMessage()) {
+class SkyflowError(val skyflowErrorCode: SkyflowErrorCode = SkyflowErrorCode.UNKNOWN_ERROR, val tag : String? = "", logLevel: LogLevel? = null, params: Array<String?> = arrayOf()) : Exception(skyflowErrorCode.getMessage()) {
 
     override var message = ""
+    internal var internalMessage = ""
     private var code = skyflowErrorCode.getCode()
 
     init {
-        this.message =  Utils.constructMessage(skyflowErrorCode.getMessage(), *params)
-        Logger.error(tag, message, logLevel)
+        val logMessage =  Utils.constructMessage(skyflowErrorCode.getMessage(), *params)
+        if(logLevel != null)
+            Logger.error(tag, logMessage, logLevel)
+        this.internalMessage = logMessage
+        this.message = "Interface : $tag - $logMessage"
     }
-
-
     fun setErrorCode(code:Int)
     {
         this.code = code
@@ -24,21 +26,12 @@ class SkyflowError(val skyflowErrorCode: SkyflowErrorCode = SkyflowErrorCode.UNK
     {
         return this.code
     }
-
-//    fun setErrorMessage(message:String)
-//    {
-//        this.message = message
-//    }
-
     fun getErrorMessage() :String
     {
         return this.message
     }
-
-//    fun setErrorResponse(vararg params: String?)
-//    {
-//        if(!params.isEmpty())
-//           message =  Utils.constructMessage(this.skyflowErrorCode.message,*params)
-//    }
+    internal fun getInternalErrorMessage():String{
+        return this.internalMessage
+    }
 }