+++ Skyflow/src/main/kotlin/Skyflow/TextField.kt
@@ -16,15 +16,22 @@ import android.widget.EditText
 import android.widget.TextView
 import androidx.annotation.RequiresApi
 import androidx.core.content.res.ResourcesCompat
-import com.Skyflow.collect.elements.validations.SkyflowValidationError
-import com.Skyflow.collect.elements.validations.ValidationSet
-import com.Skyflow.collect.elements.validations.SkyflowValidator
 import Skyflow.core.elements.state.StateforText
 import Skyflow.utils.EventName
 import android.graphics.Typeface
+import android.text.Spanned
+import androidx.core.text.isDigitsOnly
 import com.skyflow_android.R
 import org.json.JSONObject
 import kotlin.String
+import android.text.InputFilter
+import android.text.InputFilter.LengthFilter
+import android.util.Log
+import android.util.Log.w
+import com.Skyflow.collect.elements.validations.*
+import com.Skyflow.collect.elements.validations.SkyflowValidator
+import com.Skyflow.collect.elements.validations.SkyflowValidateExpireDate
+import java.util.*
 
 
 @Suppress("DEPRECATION")
@@ -33,7 +40,7 @@ class TextField @JvmOverloads constructor(
     val optionsForLogging: Options,
     attrs: AttributeSet? = null,
     defStyleAttr: Int = 0,
-) : Skyflow.Element(context, attrs, defStyleAttr){
+) : Skyflow.Element(context, attrs, defStyleAttr),BaseElement{
 
     internal var label = TextView(context)
     internal var inputField = EditText(context)
@@ -48,23 +55,35 @@ class TextField @JvmOverloads constructor(
     internal var userOnFocusListener: ((JSONObject) -> Unit)? = null
     internal var userOnBlurListener: ((JSONObject) -> Unit)? = null
     internal var userOnReadyListener: ((JSONObject) -> Unit)? = null
+    internal var expiryDateFormat = "mm/yy"
+    private var userError : String = ""
+    private  val tag = TextField::class.qualifiedName
     override fun getValue() : String {
         return actualValue
     }
 
     override fun validate() : SkyflowValidationError {
-        val str = inputField.text.toString()
-        val builtinValidationError = SkyflowValidator.validate(str,validationRules)
-        if(builtinValidationError.equals(""))
+        val str = getValue()
+        if(userError.isNotEmpty()){
+            return userError
+        }
+        var builtinValidationError = ""
+        if(isRequired && str.isEmpty()){
+            builtinValidationError = "value is empty\n"
+            setErrorText("value is required")
+            return builtinValidationError
+        }
+        builtinValidationError += SkyflowValidator.validate(str,validationRules)
+        if(builtinValidationError == "")
         {
             return if(collectInput.validations.rules.isEmpty())
                 ""
             else {
                 val customValidationError = SkyflowValidator.validate(str,collectInput.validations)
-                if(customValidationError.equals(""))
+                if(customValidationError == "")
                     ""
                 else {
-                    setError(customValidationError)
+                    setErrorText(customValidationError)
                     customValidationError
                 }
             }
@@ -72,9 +91,9 @@ class TextField @JvmOverloads constructor(
         else
         {
             if(collectInput.label.isEmpty())
-                setError("invalid field")
+                setErrorText("invalid field")
             else
-                setError("invalid "+ collectInput.label)
+                setErrorText("invalid "+ collectInput.label)
             return builtinValidationError
         }
     }
@@ -98,7 +117,7 @@ class TextField @JvmOverloads constructor(
         label.textSize = 16F
         label.setPadding(labelPadding.left,labelPadding.top,labelPadding.right,labelPadding.bottom)
         label.setTextColor(collectInput.labelStyles.base.textColor)
-        if(!collectInput.labelStyles.base.font.equals(Typeface.NORMAL))
+        if(collectInput.labelStyles.base.font != Typeface.NORMAL)
             label.typeface = ResourcesCompat.getFont(context,collectInput.labelStyles.base.font)
         label.gravity = collectInput.labelStyles.base.textAlignment
 
@@ -106,12 +125,6 @@ class TextField @JvmOverloads constructor(
 
     private fun buildTextField()
     {
-        if(collectInput.altText.isNotEmpty() || collectInput.altText != "")
-        {
-            inputField.setText(collectInput.altText)
-        }
-        else
-            inputField.setText("")
         state = StateforText(this)
         border.setColor(Color.WHITE)
         border.setStroke(collectInput.inputStyles.base.borderWidth,collectInput.inputStyles.base.borderColor)
@@ -121,11 +134,33 @@ class TextField @JvmOverloads constructor(
         inputField.gravity = collectInput.inputStyles.base.textAlignment
         inputField.hint = collectInput.placeholder
         inputField.setTextColor(collectInput.inputStyles.base.textColor)
+        inputField.inputType = fieldType.getType().keyboardType
+
         if(collectInput.inputStyles.base.font != Typeface.NORMAL)
             inputField.typeface = ResourcesCompat.getFont(context,collectInput.inputStyles.base.font)
 
-        changeCardIcon()
-        inputField.inputType = fieldType.getType().keyboardType
+        if(fieldType.equals(SkyflowElementType.EXPIRATION_DATE))
+        {
+            changeExpireDateValidations()
+        }
+        formatPatternForField(inputField.editableText)
+    }
+
+    private fun changeExpireDateValidations() {
+        validationRules.rules.clear()
+        val expireDateList = mutableListOf<String>("mm/yy","mm/yyyy","yy/mm","yyyy/mm")
+        if(expireDateList.contains(options.format.toLowerCase()))
+        {
+            expiryDateFormat = options.format.toLowerCase()
+            validationRules.add(SkyflowValidateExpireDate(format = expiryDateFormat))
+
+        }
+        else {
+            Log.w(tag,"invalid format for EXPIRATION_DATE")
+            Log.w(tag,"Using default format mm/yy for EXPIRATION_DATE")
+            validationRules.add(SkyflowValidateExpireDate(format = expiryDateFormat))
+        }
+
     }
 
     private fun buildError()
@@ -160,6 +195,10 @@ class TextField @JvmOverloads constructor(
         addView(label)
         addView(inputField)
         addView(error)
+//        error.visibility = INVISIBLE
+        if(userError.isNotEmpty()){
+            invalidTextField()
+        }
     }
 
     private fun setListenersForText() {
@@ -176,7 +215,7 @@ class TextField @JvmOverloads constructor(
 
             override fun afterTextChanged(s: Editable?) {
                 actualValue = inputField.text.toString()
-                changeCardIcon()
+                formatPatternForField(s)
                 state = StateforText(this@TextField)
                 if(userOnchangeListener !== null)
                     userOnchangeListener?.let { it((state as StateforText).getState(optionsForLogging.env)) }
@@ -193,14 +232,10 @@ class TextField @JvmOverloads constructor(
         }.also { inputField.onFocusChangeListener = it }
 
     }
-    internal fun changeCardIcon()
+    private fun changeCardIcon(cardtype:CardType)
     {
-        if(fieldType.equals(SkyflowElementType.CARD_NUMBER) && options.enableCardIcon)
-        {
-            val cardtype = CardType.forCardNumber(inputField.text.toString())
             inputField.setCompoundDrawablesRelativeWithIntrinsicBounds(cardtype.image, 0, 0, 0)
             inputField.compoundDrawablePadding = 8
-        }
     }
     private fun onFocusTextField()
     {
@@ -237,7 +272,7 @@ class TextField @JvmOverloads constructor(
         label.gravity = collectInput.labelStyles.base.textAlignment
 
         val internalState = this.state.getInternalState()
-        if(internalState["isEmpty"] as Boolean) {
+        if(internalState["isEmpty"] as Boolean && !isRequired) {
             emptyTextField()
         } else if(!(internalState["isValid"] as Boolean)) {
             invalidTextField()
@@ -250,6 +285,7 @@ class TextField @JvmOverloads constructor(
 
     private fun validTextField()
     {
+        error.visibility = INVISIBLE
         val inputFieldPadding = collectInput.inputStyles.complete.padding
         inputField.setPadding(inputFieldPadding.left,inputFieldPadding.top,inputFieldPadding.right,inputFieldPadding.bottom)
         inputField.setTextColor(collectInput.inputStyles.complete.textColor)
@@ -261,7 +297,7 @@ class TextField @JvmOverloads constructor(
             inputField.typeface = ResourcesCompat.getFont(context,collectInput.inputStyles.complete.font)
     }
 
-    private fun invalidTextField()
+    internal fun invalidTextField()
     {
         val inputFieldPadding = collectInput.inputStyles.invalid.padding
         inputField.setPadding(inputFieldPadding.left,inputFieldPadding.top,inputFieldPadding.right,inputFieldPadding.bottom)
@@ -288,7 +324,61 @@ class TextField @JvmOverloads constructor(
         if(collectInput.inputStyles.empty.font != Typeface.NORMAL)
             inputField.typeface = ResourcesCompat.getFont(context,collectInput.inputStyles.empty.font)
     }
-    internal fun setError(error: String)
+
+    var previousLength = 0
+    private fun addSlashspanToExpiryDate(editable: Editable?, expiryDateFormat: String) {
+
+        val filterArray = arrayOfNulls<InputFilter>(1)
+        filterArray[0] = LengthFilter(expiryDateFormat.length)
+        inputField.setFilters(filterArray)
+        val index = expiryDateFormat.indexOf("/")
+        val length = editable!!.length
+        if (index == length && previousLength<length) {
+            if(inputField.text.isDigitsOnly())
+            {
+                inputField.append("/")
+                previousLength = length+1
+            }
+        }
+        else if(index == length && previousLength>length)
+        {
+            inputField.setText(inputField.text.toString().subSequence(0,inputField.length()-1))
+            previousLength = length -1
+            inputField.setSelection(inputField.length())
+        }
+    }
+
+    private fun addSpaceSpanToCardNumber(editable: Editable?, spaceIndices: IntArray) {
+        val length = editable!!.length
+
+        for (index in spaceIndices) {
+            if (index <= length) {
+                editable.setSpan(
+                    Spacespan(), index - 1, index,
+                    Spanned.SPAN_EXCLUSIVE_EXCLUSIVE
+                )
+            }
+        }
+    }
+    private fun formatPatternForField(s: Editable?)
+    {
+        if(fieldType.equals(SkyflowElementType.CARD_NUMBER))
+        {
+            val cardtype = CardType.forCardNumber(inputField.text.toString().replace(" ",  "").replace("-",""))
+            if(options.enableCardIcon)
+                changeCardIcon(cardtype)
+            val filterArray = arrayOfNulls<InputFilter>(1)
+            filterArray[0] = LengthFilter(cardtype.cardLength.get(cardtype.cardLength.size-1))
+            inputField.setFilters(filterArray)
+            addSpaceSpanToCardNumber(s,cardtype.formatPattern)
+        }
+        else if(fieldType.equals(SkyflowElementType.EXPIRATION_DATE))
+        {
+            addSlashspanToExpiryDate(s,expiryDateFormat)
+        }
+    }
+
+    internal fun setErrorText(error: String)
     {
         this.error.text = error
     }
@@ -299,6 +389,55 @@ class TextField @JvmOverloads constructor(
         buildLabel()
         error.visibility = View.INVISIBLE
         actualValue = ""
+    }
+
+    override fun setError(error: String) {
+        this.userError = error
+        setErrorText(userError)
+        state = StateforText(this@TextField)
+    }
+
+    override fun resetError() {
+        this.userError = ""
+        state = StateforText(this)
+        val internalState = state.getInternalState()
+        setErrorText(internalState["validationError"].toString())
+        validTextField()
+    }
+
+    internal fun setText(value:String)
+    {
+        this.inputField.setText(value)
+    }
+
+    internal fun getErrorText():String
+    {
+        return error.text.toString()
+    }
+
+    fun setValue(value:String){
+        if(optionsForLogging.env == Env.DEV)
+        {
+            actualValue = value
+            setText(value)
+        }
+        else
+        {
+            Log.w(tag,"setValue can be called only in dev mode")
+        }
+    }
+
+    fun clearValue()
+    {
+        if(optionsForLogging.env == Env.DEV)
+        {
+            actualValue = ""
+            setText("")
+        }
+        else
+        {
+            Log.w(tag,"clearValue can be called only in dev mode")
+        }
     }
 
 }