+++ Skyflow/src/main/kotlin/Skyflow/reveal/RevealValueCallback.kt
@@ -64,26 +64,8 @@ class RevealValueCallback(var callback: Callback, var revealElements: MutableLis
             while (i < errorArray.length()) {
                 val recordObj = errorArray[i] as JSONObject
                 val tokenId = recordObj.get("token").toString()
-
-                elementsMap[tokenId]!!.error.text = "invalid token"
-                if(elementsMap[tokenId]!!.revealInput.inputStyles.invalid.font != Typeface.NORMAL)
-                    elementsMap[tokenId]!!.placeholder.typeface =
-                        ResourcesCompat.getFont(elementsMap[tokenId]!!.context,
-                            elementsMap[tokenId]!!.revealInput.inputStyles.invalid.font)
-                elementsMap[tokenId]!!.placeholder.gravity =
-                    elementsMap[tokenId]!!.revealInput.inputStyles.invalid.textAlignment
-                val padding = elementsMap[tokenId]!!.revealInput.inputStyles.invalid.padding
-                elementsMap[tokenId]!!.placeholder.setPadding(padding.left,
-                    padding.top,
-                    padding.right,
-                    padding.bottom)
-                elementsMap[tokenId]!!.placeholder.setTextColor(elementsMap[tokenId]!!.revealInput.inputStyles.invalid.textColor)
-                elementsMap[tokenId]!!.border.setStroke(elementsMap[tokenId]!!.revealInput.inputStyles.invalid.borderWidth,
-                    elementsMap[tokenId]!!.revealInput.inputStyles.invalid.borderColor)
-                elementsMap[tokenId]!!.border.cornerRadius =
-                    elementsMap[tokenId]!!.revealInput.inputStyles.invalid.cornerRadius
-                elementsMap[tokenId]!!.placeholder.setBackgroundDrawable(elementsMap[tokenId]!!.border)
-
+                elementsMap[tokenId]!!.setErrorText("invalid token")
+                elementsMap[tokenId]!!.showError()
                 i++
             }
         }