+++ Skyflow/src/test/java/com/Skyflow/CollectTest.kt
@@ -3,6 +3,7 @@ package com.Skyflow
 import Skyflow.*
 import Skyflow.collect.client.CollectAPICallback
 import Skyflow.collect.client.CollectRequestBody
+import Skyflow.collect.elements.validations.ElementValueMatchRule
 import Skyflow.core.APIClient
 import Skyflow.core.elements.state.StateforText
 import Skyflow.utils.EventName
@@ -19,6 +20,7 @@ import org.robolectric.RobolectricTestRunner
 import org.robolectric.android.controller.ActivityController
 import java.io.IOException
 import android.util.Log
+import com.Skyflow.collect.elements.validations.ValidationSet
 import junit.framework.Assert
 import junit.framework.Assert.*
 import junit.framework.TestCase
@@ -62,7 +64,7 @@ class CollectTest {
         card_number!!.inputField.setText("4111 1111 1111 1111")
 
         val state = StateforText(card_number).getInternalState()
-        Assert.assertTrue(state["isValid"] as Boolean)
+        assertTrue(state["isValid"] as Boolean)
 
     }
 
@@ -76,18 +78,28 @@ class CollectTest {
         val pin = container.create(activity,collectInput) as? TextField
         pin!!.inputField.setText("4111111")
         var state = StateforText(pin).getInternalState()
-        Assert.assertTrue(state["isValid"] as Boolean)
-
-        pin.inputField.setText("411")
-        state = StateforText(pin).getInternalState()
-        Assert.assertFalse(state["isValid"] as Boolean)
-
-
-        pin.inputField.setText("xyzzz")
-        state = StateforText(pin).getInternalState()
-        Assert.assertFalse(state["isValid"] as Boolean) //accepts only numbers
-
+        assertTrue(state["isValid"] as Boolean)
     }
+
+//    @Test
+//    fun testInvalidValueForPIN()
+//    {
+//        val container = skyflow.container(ContainerType.COLLECT)
+//        val collectInput = CollectElementInput("cards","PIN",
+//            SkyflowElementType.PIN,placeholder = "pin"
+//        )
+//        val pin = container.create(activity,collectInput) as? TextField
+//
+//        pin!!.inputField.setText("411")
+//        var state = StateforText(pin).getInternalState()
+//        assertFalse(state["isValid"] as Boolean)
+//
+//
+//        pin.inputField.setText("xyzzz")
+//        state = StateforText(pin).getInternalState()
+//        Assert.assertFalse(state["isValid"] as Boolean) //accepts only numbers
+//
+//    }
 
     @Test
     fun testEmptyStateForSkyflowElement()
@@ -121,24 +133,6 @@ class CollectTest {
 
         val state = StateforText(card_number!!).getInternalState()
         Assert.assertTrue(state["isEmpty"] as Boolean)
-    }
-
-    @Test
-    fun testInvalidValueSkyflowElement()
-    {
-        val container = skyflow.container(ContainerType.COLLECT)
-        val collectInput = CollectElementInput("cards","card_number",
-            SkyflowElementType.CARD_NUMBER,placeholder = "card number"
-        )
-        val card_number = container.create(activity,collectInput) as? TextField
-        card_number!!.inputField.setText("4111")
-        card_number.state = StateforText(card_number)
-        val state = StateforText(card_number)
-        Assert.assertFalse(state.getInternalState()["isValid"] as Boolean)
-        assertNotNull(card_number.state.show())
-        assertEquals(state.getInternalState().get("inputLength"),4)
-        assertEquals(state.getState(Env.DEV).get("value"),"") // value will change when text event afterTextChanged triggers
-
     }
 
     @Test
@@ -179,7 +173,7 @@ class CollectTest {
 
             override fun onFailure(exception: Any) {
                 val skyflowError = SkyflowError(SkyflowErrorCode.ELEMENT_NOT_MOUNTED,params = arrayOf(card_number.columnName))
-                assertEquals((exception as SkyflowError).message.trim(),skyflowError.getErrorMessage().trim())
+                assertEquals((exception as SkyflowError).getInternalErrorMessage().trim(),skyflowError.getInternalErrorMessage().trim())
                 assertEquals(400,skyflowError.getErrorcode())
             }
         })
@@ -197,7 +191,8 @@ class CollectTest {
             val cvv : TextField = container.create(activity,collectInput2, options)
 
             card_number.inputField.setText("4111 1111 1111 1111")
-            cvv.inputField.setText("2")
+            card_number.actualValue = "4111 1111 1111 1111"
+            cvv.inputField.setText("142")
             activity.addContentView(card_number,layoutParams)
             activity.addContentView(cvv,layoutParams)
             card_number.state = StateforText(card_number)
@@ -210,8 +205,8 @@ class CollectTest {
                 }
 
                 override fun onFailure(exception: Any) {
-                    val skyflowError = SkyflowError(SkyflowErrorCode.INVALID_INPUT,params = arrayOf("for cvv INVALID_LENGTH_MATCH"))
-                    assertEquals((exception as SkyflowError).message.trim(),skyflowError.getErrorMessage().trim())
+                    val skyflowError = SkyflowError(SkyflowErrorCode.INVALID_INPUT,params = arrayOf("for cvv value is empty"))
+                    assertEquals((exception as SkyflowError).getInternalErrorMessage().trim(),skyflowError.getInternalErrorMessage().trim())
                 }
             })
         }
@@ -229,6 +224,9 @@ class CollectTest {
         val collectInput2 = CollectElementInput("cards","cvv",SkyflowElementType.CVV, label = "cvv")
         val card_number = container.create(activity,collectInput1, options)
         val cvv = container.create(activity,collectInput2,options)
+        cvv.inputField.setText("123")
+        cvv.actualValue = "123"
+        cvv.state = StateforText(cvv)
         activity.addContentView(card_number,layoutParams)
         activity.addContentView(cvv,layoutParams)
 
@@ -239,9 +237,8 @@ class CollectTest {
             }
 
             override fun onFailure(exception: Any) {
-                val skyflowError = SkyflowError(SkyflowErrorCode.INVALID_INPUT,params = arrayOf("card_number is empty\n" +
-                        "cvv is empty"))
-                assertEquals((exception as SkyflowError).message.trim(),skyflowError.getErrorMessage().trim())
+                val skyflowError = SkyflowError(SkyflowErrorCode.INVALID_INPUT,params = arrayOf("for card_number value is empty"))
+                assertEquals((exception as SkyflowError).getInternalErrorMessage().trim(),skyflowError.getInternalErrorMessage().trim())
             }
         })
     }
@@ -268,7 +265,8 @@ class CollectTest {
             override fun onSuccess(responseBody: Any) {
             }
             override fun onFailure(exception: Any) {
-                assertEquals((exception as SkyflowError).message,SkyflowErrorCode.EMPTY_VAULT_ID.getMessage())
+                val skyflowError = SkyflowError(SkyflowErrorCode.EMPTY_VAULT_ID)
+                assertEquals((exception as SkyflowError).getInternalErrorMessage(),skyflowError.getInternalErrorMessage())
             }
         })
 
@@ -304,7 +302,8 @@ class CollectTest {
             }
 
             override fun onFailure(exception: Any) {
-                assertEquals((exception as SkyflowError).message,SkyflowErrorCode.EMPTY_VAULT_URL.getMessage())
+                val skyflowError = SkyflowError(SkyflowErrorCode.EMPTY_VAULT_URL)
+                assertEquals((exception as SkyflowError).getInternalErrorMessage(),skyflowError.getInternalErrorMessage())
             }
 
 
@@ -343,7 +342,7 @@ class CollectTest {
 
             override fun onFailure(exception: Any) {
                 val skyflowError = SkyflowError(SkyflowErrorCode.INVALID_VAULT_URL,params = arrayOf(configuration.vaultURL))
-                assertEquals((exception as SkyflowError).message,skyflowError.getErrorMessage())
+                assertEquals((exception as SkyflowError).getInternalErrorMessage(),skyflowError.getInternalErrorMessage())
             }
 
 
@@ -368,7 +367,8 @@ class CollectTest {
             }
 
             override fun onFailure(exception: Any) {
-                assertEquals((exception as SkyflowError).message,SkyflowErrorCode.MISSING_TABLE.getMessage())
+                val skyflowError = SkyflowError(SkyflowErrorCode.MISSING_TABLE_IN_ELEMENT, params = arrayOf(card_number.fieldType.toString()))
+                assertEquals((exception as SkyflowError).getInternalErrorMessage(),skyflowError.getInternalErrorMessage())
             }
 
 
@@ -394,8 +394,9 @@ class CollectTest {
             }
 
             override fun onFailure(exception: Any) {
-                Log.d("exc",(exception as SkyflowError).message)
-                assertEquals((exception as SkyflowError).message,SkyflowErrorCode.EMPTY_TABLE_NAME.getMessage())
+                Log.d("exc",(exception as SkyflowError).getInternalErrorMessage())
+                val skyflowError = SkyflowError(SkyflowErrorCode.ELEMENT_EMPTY_TABLE_NAME, params = arrayOf(card_number.fieldType.toString()))
+                assertEquals((exception as SkyflowError).getInternalErrorMessage(),skyflowError.getInternalErrorMessage())
             }
 
 
@@ -420,7 +421,8 @@ class CollectTest {
             }
 
             override fun onFailure(exception: Any) {
-                assertEquals((exception as SkyflowError).message,SkyflowErrorCode.MISSING_COLUMN.getMessage())
+                val skyflowError = SkyflowError(SkyflowErrorCode.MISSING_COLUMN, params = arrayOf(card_number.fieldType.toString()))
+                assertEquals((exception as SkyflowError).getInternalErrorMessage(),skyflowError.getInternalErrorMessage())
             }
 
 
@@ -448,7 +450,8 @@ class CollectTest {
             }
 
             override fun onFailure(exception: Any) {
-                assertEquals((exception as SkyflowError).message,SkyflowErrorCode.EMPTY_COLUMN_NAME.getMessage())
+                val skyflowError = SkyflowError(SkyflowErrorCode.EMPTY_COLUMN_NAME, params = arrayOf(card_number.fieldType.toString()))
+                assertEquals((exception as SkyflowError).getInternalErrorMessage(),skyflowError.getInternalErrorMessage())
             }
 
 
@@ -498,7 +501,7 @@ class CollectTest {
         card_number!!.inputField.setText("4111 1111 1111 1111")
         activity.addContentView(card_number,layoutParams)
         assertEquals(2,card_number.collectInput.inputStyles.base.borderWidth)
-        card_number.setError("error occured")
+        card_number.setErrorText("error occured")
         assertTrue(card_number.error.text.toString().equals("error occured"))
         card_number.unmount()
         assertTrue(card_number.actualValue.isEmpty())
@@ -681,7 +684,7 @@ class CollectTest {
 
             override fun onFailure(exception: Any) {
                 val skyflowError = SkyflowError(SkyflowErrorCode.INVALID_VAULT_URL,params = arrayOf(apiClient.vaultURL))
-                TestCase.assertEquals(skyflowError.getErrorMessage(),(exception as Exception).message.toString())
+                TestCase.assertEquals(skyflowError.getInternalErrorMessage(),(exception as Exception).message.toString())
             }
 
         }, InsertOptions(),LogLevel.ERROR)
@@ -793,8 +796,8 @@ class CollectTest {
 
             override fun onFailure(exception: Any) {
                 val skyflowError = SkyflowError(SkyflowErrorCode.DUPLICATE_COLUMN_FOUND,params = arrayOf(collectInput.table,collectInput.column))
-                TestCase.assertEquals(skyflowError.getErrorMessage(),
-                    (exception as SkyflowError).getErrorMessage())
+                TestCase.assertEquals(skyflowError.getInternalErrorMessage(),
+                    (exception as SkyflowError).getInternalErrorMessage())
             }
 
         },LogLevel.ERROR)
@@ -831,9 +834,9 @@ class CollectTest {
             }
 
             override fun onFailure(exception: Any) {
-                val skyflowError = SkyflowError(SkyflowErrorCode.EMPTY_TABLE_NAME)
-                TestCase.assertEquals(skyflowError.getErrorMessage(),
-                    (exception as SkyflowError).getErrorMessage())
+                val skyflowError = SkyflowError(SkyflowErrorCode.EMPTY_TABLE_KEY)
+                TestCase.assertEquals(skyflowError.getInternalErrorMessage(),
+                    (exception as SkyflowError).getInternalErrorMessage())
             }
 
         },LogLevel.ERROR)
@@ -870,9 +873,9 @@ class CollectTest {
             }
 
             override fun onFailure(exception: Any) {
-                val skyflowError = SkyflowError(SkyflowErrorCode.MISSING_TABLE)
-                TestCase.assertEquals(skyflowError.getErrorMessage(),
-                    (exception as SkyflowError).getErrorMessage())
+                val skyflowError = SkyflowError(SkyflowErrorCode.MISSING_TABLE_KEY)
+                TestCase.assertEquals(skyflowError.getInternalErrorMessage(),
+                    (exception as SkyflowError).getInternalErrorMessage())
             }
 
         },LogLevel.ERROR)
@@ -911,8 +914,8 @@ class CollectTest {
 
             override fun onFailure(exception: Any) {
                 val skyflowError = SkyflowError(SkyflowErrorCode.INVALID_TABLE_NAME)
-                TestCase.assertEquals(skyflowError.getErrorMessage(),
-                    (exception as SkyflowError).getErrorMessage())
+                TestCase.assertEquals(skyflowError.getInternalErrorMessage(),
+                    (exception as SkyflowError).getInternalErrorMessage())
             }
 
         },LogLevel.ERROR)
@@ -950,52 +953,13 @@ class CollectTest {
 
             override fun onFailure(exception: Any) {
                 val skyflowError = SkyflowError(SkyflowErrorCode.FIELDS_KEY_ERROR)
-                TestCase.assertEquals(skyflowError.getErrorMessage(),
-                    (exception as SkyflowError).getErrorMessage())
+                TestCase.assertEquals(skyflowError.getInternalErrorMessage(),
+                    (exception as SkyflowError).getInternalErrorMessage())
             }
 
         },LogLevel.ERROR)
     }
 
-
-    @Test
-    fun testMissingColumnInAdditionalFields()
-    {
-        val container = skyflow.container(ContainerType.COLLECT)
-        val collectInput = CollectElementInput("cards","cvv",
-            SkyflowElementType.CVV,placeholder = "cvv"
-        )
-        val collectInput1 = CollectElementInput("cards","card_number",
-            SkyflowElementType.CARD_NUMBER,placeholder = "card number"
-        )
-        val cvv = container.create(activity,collectInput,CollectElementOptions())
-        val card_number = container.create(activity,collectInput1,CollectElementOptions())
-
-        val records = JSONObject()
-        val recordsArray = JSONArray()
-        val record = JSONObject()
-        val fields = JSONObject()
-        record.put("table", "cards")
-        fields.put("", "41111111111")
-        fields.put("expiry_date","11/22")
-        record.put("fields", fields)
-        recordsArray.put(record)
-        records.put("records", recordsArray)
-
-        CollectRequestBody.createRequestBody(container.elements, records,object : Callback
-        {
-            override fun onSuccess(responseBody: Any) {
-
-            }
-
-            override fun onFailure(exception: Any) {
-                val skyflowError = SkyflowError(SkyflowErrorCode.EMPTY_COLUMN_NAME)
-                TestCase.assertEquals(skyflowError.getErrorMessage(),
-                    (exception as SkyflowError).getErrorMessage())
-            }
-
-        },LogLevel.ERROR)
-    }
 
     @Test
     fun testEmptyFieldsInAdditionalFields()
@@ -1027,8 +991,8 @@ class CollectTest {
 
             override fun onFailure(exception: Any) {
                 val skyflowError = SkyflowError(SkyflowErrorCode.EMPTY_FIELDS)
-                TestCase.assertEquals(skyflowError.getErrorMessage(),
-                    (exception as SkyflowError).getErrorMessage())
+                TestCase.assertEquals(skyflowError.getInternalErrorMessage(),
+                    (exception as SkyflowError).getInternalErrorMessage())
             }
 
         },LogLevel.ERROR)
@@ -1059,8 +1023,8 @@ class CollectTest {
 
             override fun onFailure(exception: Any) {
                 val skyflowError = SkyflowError(SkyflowErrorCode.EMPTY_RECORDS)
-                TestCase.assertEquals(skyflowError.getErrorMessage(),
-                    (exception as SkyflowError).getErrorMessage())
+                TestCase.assertEquals(skyflowError.getInternalErrorMessage(),
+                    (exception as SkyflowError).getInternalErrorMessage())
             }
 
         },LogLevel.ERROR)
@@ -1090,8 +1054,8 @@ class CollectTest {
 
             override fun onFailure(exception: Any) {
                 val skyflowError = SkyflowError(SkyflowErrorCode.INVALID_RECORDS)
-                TestCase.assertEquals(skyflowError.getErrorMessage(),
-                    (exception as SkyflowError).getErrorMessage())
+                TestCase.assertEquals(skyflowError.getInternalErrorMessage(),
+                    (exception as SkyflowError).getInternalErrorMessage())
             }
 
         },LogLevel.ERROR)
@@ -1129,14 +1093,145 @@ class CollectTest {
 
             override fun onFailure(exception: Any) {
                 val skyflowError = SkyflowError(SkyflowErrorCode.DUPLICATE_COLUMN_FOUND,params = arrayOf(collectInput1.table,collectInput1.column))
-                TestCase.assertEquals(skyflowError.getErrorMessage(),
-                    (exception as SkyflowError).getErrorMessage())
+                TestCase.assertEquals(skyflowError.getInternalErrorMessage(),
+                    (exception as SkyflowError).getInternalErrorMessage())
             }
 
         },LogLevel.ERROR)
     }
 
     //end collect
+
+    @Test
+    fun testmmyyExpireDate()
+    {
+        val container = skyflow.container(ContainerType.COLLECT)
+        val collectInput = CollectElementInput("cards","expire date",
+            SkyflowElementType.EXPIRATION_DATE,placeholder = "expire date"
+        )
+        val expireDate = container.create(activity,collectInput,CollectElementOptions(format = "mm/yy"))
+        activity.addContentView(expireDate,layoutParams)
+        expireDate.inputField.setText("12/39")
+        expireDate.state = StateforText(expireDate)
+        assertTrue(expireDate.state.getInternalState().get("isValid") as Boolean)
+
+        expireDate.inputField.setText("12/20")
+        expireDate.state = StateforText(expireDate)
+        assertFalse(expireDate.state.getInternalState().get("isValid") as Boolean)
+
+    }
+
+    @Test
+    fun testmmyyyyExpireDate()
+    {
+        val container = skyflow.container(ContainerType.COLLECT)
+        val collectInput = CollectElementInput("cards","expire date",
+            SkyflowElementType.EXPIRATION_DATE,placeholder = "expire date"
+        )
+        val expireDate = container.create(activity,collectInput,CollectElementOptions(format = "mm/yyyy"))
+        activity.addContentView(expireDate,layoutParams)
+        expireDate.inputField.setText("12/2039")
+        expireDate.state = StateforText(expireDate)
+        assertTrue(expireDate.state.getInternalState().get("isValid") as Boolean)
+
+        expireDate.inputField.setText("12/20njn")
+        expireDate.state = StateforText(expireDate)
+        assertFalse(expireDate.state.getInternalState().get("isValid") as Boolean)
+
+    }
+    @Test
+    fun testyymmExpireDate()
+    {
+        val container = skyflow.container(ContainerType.COLLECT)
+        val collectInput = CollectElementInput("cards","expire date",
+            SkyflowElementType.EXPIRATION_DATE,placeholder = "expire date"
+        )
+        val expireDate = container.create(activity,collectInput,CollectElementOptions(format = "yy/mm"))
+        activity.addContentView(expireDate,layoutParams)
+        expireDate.inputField.setText("39/12")
+        expireDate.state = StateforText(expireDate)
+        assertTrue(expireDate.state.getInternalState().get("isValid") as Boolean)
+
+        expireDate.inputField.setText("12/20")
+        expireDate.state = StateforText(expireDate)
+        assertFalse(expireDate.state.getInternalState().get("isValid") as Boolean)
+
+    }
+
+    @Test
+    fun testyyyymmExpireDate()
+    {
+        val container = skyflow.container(ContainerType.COLLECT)
+        val collectInput = CollectElementInput("cards","expire date",
+            SkyflowElementType.EXPIRATION_DATE,placeholder = "expire date"
+        )
+        val expireDate = container.create(activity,collectInput,CollectElementOptions(format = "yyyy/mm"))
+        activity.addContentView(expireDate,layoutParams)
+        expireDate.inputField.setText("2039/12")
+        expireDate.state = StateforText(expireDate)
+        assertTrue(expireDate.state.getInternalState().get("isValid") as Boolean)
+
+        expireDate.inputField.setText("12/20")
+        expireDate.state = StateforText(expireDate)
+        assertFalse(expireDate.state.getInternalState().get("isValid") as Boolean)
+
+    }
+
+    @Test
+    fun testSetAndResetError()
+    {
+        val container = skyflow.container(ContainerType.COLLECT)
+        val collectInput = CollectElementInput("cards","expire date",
+            SkyflowElementType.EXPIRATION_DATE,placeholder = "expire date"
+        )
+        val expireDate = container.create(activity,collectInput,CollectElementOptions(format = "yyyy/mm"))
+        activity.addContentView(expireDate,layoutParams)
+        expireDate.inputField.setText("209/12")
+        expireDate.state = StateforText(expireDate)
+        expireDate.setError("custom error")
+        assertEquals("custom error",expireDate.getErrorText())
+
+        expireDate.resetError()
+        assertEquals("INVALID_EXPIRE_DATE",expireDate.getErrorText())
+
+
+    }
+
+
+
+
+    @Test
+    fun testValidElementMatchRuleWithDuplicateTableAndColumn() //valid
+    {
+        val container = skyflow.container(ContainerType.COLLECT)
+        val collectInput = CollectElementInput("cards","pin",
+            SkyflowElementType.EXPIRATION_DATE,placeholder = "enter pin"
+        )
+        val pin = container.create(activity,collectInput)
+        val validationSet = ValidationSet()
+        validationSet.add(ElementValueMatchRule(pin,"not matched with pin"))
+        val collectInput1 = CollectElementInput("cards","pin",
+            SkyflowElementType.EXPIRATION_DATE,placeholder = "confirm pin", validations = validationSet
+        )
+        val confirmPin = container.create(activity,collectInput1)
+
+        activity.addContentView(pin,layoutParams)
+        activity.addContentView(confirmPin,layoutParams)
+
+        container.collect(object : Callback
+        {
+            override fun onSuccess(responseBody: Any) {
+
+            }
+
+            override fun onFailure(exception: Any) {
+                assertEquals("no error",(exception as Exception).message.toString())
+            }
+
+        })
+
+
+    }